jwt added

+++ controller/commentController.js
@@ -1,13 +1,20 @@
 const Comment= require('../models/commentsSchema')
 const Ticket = require('../models/TicketSchema')
+
 exports.addComment = async(req,res)=>{
     try {
-       const {commentBy,postId,comment} = req.body;
-       const dbComment =  await Comment.create({
-        commentBy,postId,comment
-       })
-       await Ticket.findByIdAndUpdate(postId,{$push:{comments:dbComment.id}})
-       res.json(dbComment)
+        if(req.user){
+            const {id} = req.user
+            const {postId,comment} = req.body;
+            const dbComment =  await Comment.create({
+             commentBy:id,postId,comment
+            })
+            await Ticket.findByIdAndUpdate(postId,{$push:{comments:dbComment.id}})
+            res.json(dbComment)
+        }else{
+            throw new Error('User not Authorised')
+        }
+       
     } catch (error) {
         throw new Error(error)
     }
@@ -17,12 +24,15 @@ exports.addComment = async(req,res)=>{
 
 exports.deleteComment=async(req,res)=>{
     try {
-        const {id,postId} = req.body
-        await Ticket.findByIdAndUpdate(postId,{$pull:{
-            comments:id
-        }})
-        await Comment.findByIdAndDelete(id)
-        res.json(true)
+        if(req.user){
+            const {id,postId} = req.body
+            await Ticket.findByIdAndUpdate(postId,{$pull:{
+                comments:id
+            }})
+            await Comment.findByIdAndDelete(id)
+            res.json(true)
+        }
+        
     } catch (error) {
         throw new Error(error)
     }
